% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appaddnewuser.R
\name{appaddnewuser}
\alias{appaddnewuser}
\title{Add a new user to the DataMuster database}
\usage{
appaddnewuser(email, username, password)
}
\arguments{
\item{email}{the email address of the user}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the user has been added successfully
}
\description{
This function adds new users to the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
