% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appalmsgrowth.R
\name{appalmsgrowth}
\alias{appalmsgrowth}
\title{Retrieves data for the ALMS Growth graph from the DataMuster database}
\usage{
appalmsgrowth(
  property,
  sex,
  category,
  paddock,
  zoom,
  start,
  timezone,
  username,
  password
)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{sex}{the sex of the cattle to be returned, determined by the "Males or Females" filter}

\item{category}{the category of cattle to be returned, determined by the "Breeders or Growers" filter}

\item{paddock}{the paddock allocation of the cattle to be returned, determined by selecting a paddock on the map}

\item{zoom}{indicates whether to return cattle from the whole property or to filter cattle by paddock, determined by the "Paddock Groups" filter}

\item{start}{the minimum date of data to be returned, determined by the "Period for ALMS graphs" filter}

\item{timezone}{the timezone of the property to display the weekly weight data}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
A dataframe of summarised data showing the average weight of cattle by date and the number of cattle included in the analysis
}
\description{
This function retreives weekly weight data from the DataMuster database and prepares the data for graphical display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
