% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appalmsusegraph.R
\name{appalmsusegraph}
\alias{appalmsusegraph}
\title{Retrieves cattle ALMS useage information from the DataMuster database}
\usage{
appalmsusegraph(
  property,
  start,
  sex,
  category,
  alms,
  timezone,
  username,
  password
)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle and a daily visit indicator, 0 = not recorded and 1 = recorded
}
\description{
This function allows cattle ALMS use data to be retreived from the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
