% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdailywtstable.R
\name{appdailywtstable}
\alias{appdailywtstable}
\title{Retrieves data from the DataMuster database for the Daily Weights table}
\usage{
appdailywtstable(
  property,
  sex,
  category,
  paddock,
  zoom,
  timezone,
  start,
  username,
  password
)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{sex}{the sex of the cattle to be returned, determined by the "Males or Females" filter}

\item{category}{the category of cattle to be returned, determined by the "Breeders or Growers" filter}

\item{paddock}{the paddock allocation of the cattle to be returned, determined by selecting a paddock on the map}

\item{zoom}{indicates whether to return cattle from the whole property or to filter cattle by paddock, determined by the "Paddock Groups" filter}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle RFID numbers, management tags and daily weight data
}
\description{
This function retreives cattle daily weight data from the DataMuster database and prepares the data for table display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
