% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeidupload.R
\name{appeidupload}
\alias{appeidupload}
\title{Upload DataMuster EID data to the DataMuster database}
\usage{
appeidupload(RFID, datetime, ALMS, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of the cattle RFID numbers}

\item{datetime}{a list of timestamp data in POSIXct format}

\item{ALMS}{a list of ALMS identification codes as character entries}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the data has been successfully uploaded
}
\description{
This function allows DataMuster EID data to be uploaded via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
