% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetcarryingcapacity.R
\name{appgetcarryingcapacity}
\alias{appgetcarryingcapacity}
\title{Retreives carrying capacity information from the DataMuster database}
\usage{
appgetcarryingcapacity(property, period, username, password)
}
\arguments{
\item{property}{a list of property names associated with the user}

\item{period}{the time period for the data required, either "12months" or "seasonal"}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing paddock names and carrying capacity information
}
\description{
This function retrieves property information from the DataMuster database for configuration of the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
