% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetcattle.R
\name{appgetcattle}
\alias{appgetcattle}
\title{Retrieves cattle from the DataMuster database}
\usage{
appgetcattle(property, sex, category, paddock, zoom, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a spatialpointsdataframe with a list of cattle RFID numbers, management tags, category and sex information, current paddock allocations and weight data
}
\description{
This function provides a list of cattle for a property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
