% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetcattlelist.R
\name{appgetcattlelist}
\alias{appgetcattlelist}
\title{Retrieves cattle from the DataMuster database}
\usage{
appgetcattlelist(property, paddock, zoom, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{paddock}{the paddock allocation of the cattle to be returned, determined by selecting a paddock on the map}

\item{zoom}{indicates whether to return cattle from the whole property or to filter cattle by paddock, determined by the "Paddock Groups" filter}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle numbers by paddock
}
\description{
This function retreives a list of cattle and their paddock allocations for display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
