% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetpaddocklist.R
\name{appgetpaddocklist}
\alias{appgetpaddocklist}
\title{Retrieves paddock information from the DataMuster database}
\usage{
appgetpaddocklist(property, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a spatialpolygonsdataframe with a list of the paddock names and associated polygon coordinates
}
\description{
This function retrieves a list of paddock polygons for display on the property map on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
