% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetstations.R
\name{appgetstations}
\alias{appgetstations}
\title{Retreive property information from the DataMuster database}
\usage{
appgetstations(property = NULL, username, password)
}
\arguments{
\item{property}{a list of property names, if NULL all active properties in the database will be returned}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing property names
}
\description{
This function retrieves property information from the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
