% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetstockingrates.R
\name{appgetstockingrates}
\alias{appgetstockingrates}
\title{Retreives property information from the DataMuster database}
\usage{
appgetstockingrates(property, start, timezone, username, password)
}
\arguments{
\item{property}{a list of property names associated with the user}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing the number of daily AE's by paddock
}
\description{
This function retrieves property information from the DataMuster database for configuration of the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
