% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetusersettings.R
\name{appgetusersettings}
\alias{appgetusersettings}
\title{Retrieves user settings from the DataMuster database}
\usage{
appgetusersettings(email, username = NULL, password = NULL)
}
\arguments{
\item{email}{the email address of the registered user}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing the information associated with the user settings
}
\description{
This function allows user settings to be retreived from the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
