% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmanagetableNEW.R
\name{appmanagetableNEW}
\alias{appmanagetableNEW}
\title{Retrieves data from the DataMuster database for the Manage Cattle table}
\usage{
appmanagetableNEW(property, sex, category, paddock, zoom, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{sex}{the sex of the cattle to be returned, determined by the "Males or Females" filter}

\item{category}{the category of cattle to be returned, determined by the "Breeders or Growers" filter}

\item{paddock}{the paddock allocation of the cattle to be returned, determined by selecting a paddock on the map}

\item{zoom}{indicates whether to return cattle from the whole property or to filter cattle by paddock, determined by the "Paddock Groups" filter}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle RFID numbers, management tags, category and sex information, current paddock allocations and the latest crush weight and date
}
\description{
This function retreives cattle data from the DataMuster database and prepares the data for table display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
