% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmovecattle.R
\name{appmovecattle}
\alias{appmovecattle}
\title{Update cattle coordinates in the DataMuster database}
\usage{
appmovecattle(property, paddock, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{paddock}{the paddock allocation of the cattle}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the data has been updated successfully
}
\description{
This function updates cattle spatial coordinates, after moving or removing cattle, for display on the property map on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
