% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appnewuser.R
\name{appnewuser}
\alias{appnewuser}
\title{Add a new user to the DataMuster database}
\usage{
appnewuser(email, username = NULL, password = NULL)
}
\arguments{
\item{email}{the email address of the user}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the data has been successfully added
}
\description{
This function allows new users to be added to the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
