% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apppaddocks.R
\name{apppaddocks}
\alias{apppaddocks}
\title{Retrieves paddock information from the DataMuster database}
\usage{
apppaddocks(property, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a spatialpolygonsdataframe with a list of the paddock names and associated polygon coordinates
}
\description{
This function allows a list of paddock polygons to be retreived from the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
