% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appreaddata.R
\name{appreaddata}
\alias{appreaddata}
\title{Read a datafile from the DataMuster database}
\usage{
appreaddata(property, filepath, filetype, username, password)
}
\arguments{
\item{property}{the name of the property of the registered user}

\item{filepath}{the path of the file to be read, determined by navigation by the user to the file after selecting the 'Browse' button}

\item{filetype}{the type of file to be read, "DataMuster WoW (.txt)", "DataMuster EID (.txt)", "Gallagher TSi Session Data (.csv)" or "Calving Data (.csv)"}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe containing the data from the file
}
\description{
This function reads cattle data from a file and prepares the data for table display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
