% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprecorddeath.R
\name{apprecorddeath}
\alias{apprecorddeath}
\title{Record a cattle death to the DataMuster database}
\usage{
apprecorddeath(RFID, MTag, date, cause, property, paddock, username, password)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{MTag}{a list of cattle management tag numbers}

\item{date}{the date that the animal died in date format, the default is today's date}

\item{cause}{the cause of death}

\item{property}{the name of the property that the cattle belong}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not information has been successfully recorded
}
\description{
This function allows death information to be recorded for individuals or groups of cattle to the DataMuster database via the DataMuster website.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
