% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appremovecattle.R
\name{appremovecattle}
\alias{appremovecattle}
\title{Archive cattle in the DataMuster database}
\usage{
appremovecattle(RFID, MTag, property, paddock, date, username, password)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{MTag}{a list of cattle management tag numbers}

\item{property}{the name of the property that cattle belong}

\item{paddock}{the name of the paddock that the cattle belong}

\item{date}{the date that the animal was removed from the property in date format, the default is today's date}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the information has been successfully recorded
}
\description{
This function allows individuals or groups of cattle to be archived in the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
