% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptodaysdata.R
\name{apptodaysdata}
\alias{apptodaysdata}
\title{Retrieve daily WoW data from the DataMuster database via the DataMuster website app}
\usage{
apptodaysdata(alms, timezone, start, username, password)
}
\arguments{
\item{alms}{the name of the alms unit}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password}

\item{property}{the name of the property to search the database}
}
\value{
a dataframe of WoW weights that have been recorded during the specified time period
}
\description{
This function allows daily ALMS data to be retreived from the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
