% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptsisessionupload.R
\name{apptsisessionupload}
\alias{apptsisessionupload}
\title{Upload Gallagher TSi session data to the 'apptsi' collection in the DataMuster DMIoT database}
\usage{
apptsisessionupload(
  Tag,
  EID,
  date,
  liveweight,
  property,
  NLIS = NULL,
  pdkfrom = NULL,
  pdkto = NULL,
  notes = NULL,
  username,
  password
)
}
\arguments{
\item{Tag}{a list of cattle management tag numbers}

\item{EID}{a list of cattle RFID numbers}

\item{date}{a list of dates in date format}

\item{liveweight}{a list of cattle weights}

\item{property}{the name of the property that the cattle belong}

\item{NLIS}{a list of cattle NLIS numbers}

\item{pdkfrom}{a list of paddock names that the cattle have moved from}

\item{pdkto}{a list of paddock names that the cattle have moved to}

\item{notes}{observations made by the stationhand as a character entry}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the data has been successfully uploaded
}
\description{
This function uploads Gallagher TSi session data to the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
