% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appupdatepaddock.R
\name{appupdatepaddock}
\alias{appupdatepaddock}
\title{Update cattle allocations to paddocks}
\usage{
appupdatepaddock(RFID, MTag, property, paddock, date, username, password)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{MTag}{a list of cattle management tag numbers}

\item{property}{the name of the property that the cattle belong}

\item{paddock}{the name of the paddock that the cattle belong}

\item{date}{the date that the cattle were moved in date format, the default is today's date}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the information has been successfully updated
}
\description{
This function allows individuals or groups of cattle to be allocated to paddocks in the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
