% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appweeklywts.R
\name{appweeklywts}
\alias{appweeklywts}
\title{Retrieves cattle weekly ALMS weights from the DataMuster database via the DataMuster website app}
\usage{
appweeklywts(RFID, start, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{start}{a start date to be returned, this has to be in date format}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}

\item{end}{an end date to be returned, this has to be in date format}
}
\value{
A dataframe of the RFID numbers that have been returned and weekly average weights for each animal
}
\description{
This function allows cattle ALMS weekly weight data to be retreived from the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
