% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appweightstable.R
\name{appweightstable}
\alias{appweightstable}
\title{Retrieves data from the DataMuster database for the Table of Cattle Weights Data}
\usage{
appweightstable(
  property,
  sex,
  category,
  paddock,
  zoom,
  timezone,
  username,
  password
)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{sex}{the sex of the cattle to be returned, determined by the "Males or Females" filter}

\item{category}{the category of cattle to be returned, determined by the "Breeders or Growers" filter}

\item{paddock}{the paddock allocation of the cattle to be returned, determined by selecting a paddock on the map}

\item{zoom}{indicates whether to return cattle from the whole property or to filter cattle by paddock, determined by the "Paddock Groups" filter}

\item{timezone}{the timezone of the property to display the weekly weight data}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle RFID numbers, management tags, category and sex information, current paddock allocations and the latest ALMS and crush weights and dates
}
\description{
This function retreives the latest cattle weight data from the DataMuster database and prepares the data for table display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
