% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appwowdata.R
\name{appwowdata}
\alias{appwowdata}
\title{Retrieve daily WoW data from the DataMuster database via the DataMuster website app}
\usage{
appwowdata(
  start = NULL,
  end = NULL,
  location = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{start}{a start date to be returned in date format}

\item{end}{an end date to be returned in date format.}

\item{location}{the name of the alms unit, if NULL data for all units will be returned}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password}
}
\value{
a dataframe of WoW weights that have been recorded during the specified time period
}
\description{
This function allows daily ALMS data to be retreived from the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
