% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appwowupload.R
\name{appwowupload}
\alias{appwowupload}
\title{Upload DataMuster WoW data to the DataMuster database}
\usage{
appwowupload(RFID, datetime, weight, ALMS, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of the cattle RFID numbers}

\item{datetime}{a list of timestamps in POSIXct format}

\item{weight}{a list of cattle weights}

\item{ALMS}{a list of ALMS identification codes as character entries}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the data has been successfully uploaded
}
\description{
This function allows DataMuster WoW data to be uploaded to the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
