% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appalmsusegraph.R
\name{appalmsusegraph}
\alias{appalmsusegraph}
\title{Retrieves data for the ALMS Use graph from the DataMuster database}
\usage{
appalmsusegraph(property, start, sex, category, alms, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{start}{the minimum date of data to be returned, determined by the "Period for ALMS graphs" filter}

\item{sex}{the sex of the cattle to be returned, determined by the "Males or Females" filter}

\item{category}{the category of cattle to be returned, determined by the "Breeders or Growers" filter}

\item{alms}{the allocated alms unit of the cattle to be returned, determined from the "ALMS Access:" drop down selector}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe with a list of cattle and a daily visit indicator, 0 = not recorded and 1 = recorded
}
\description{
This function retreives daily weight data from the DataMuster database and prepares the data for graphical display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
