% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appgetusersettings.R
\name{appgetusersettings}
\alias{appgetusersettings}
\title{Retrieves user settings from the DataMuster database}
\usage{
appgetusersettings(email, username, password)
}
\arguments{
\item{email}{the email address of the registered user}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing the users last used language and property
}
\description{
This function retreives user website settings from the DataMuster database for configuration of the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
