% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptodaysdata.R
\name{apptodaysdata}
\alias{apptodaysdata}
\title{Retrieves daily WoW data from the DataMuster database}
\usage{
apptodaysdata(property, alms, timezone, start, username, password)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{alms}{the name of the alms unit}

\item{timezone}{the timezone of the property to display the weight data}

\item{start}{the minimum date of data to be returned}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe of WoW weights that have been recorded during the specified time period
}
\description{
This function retrieves daily ALMS data from the DataMuster for table display on the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
