% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appupdatepaddock.R
\name{appupdatepaddock}
\alias{appupdatepaddock}
\title{Update cattle allocations to paddocks}
\usage{
appupdatepaddock(
  RFID,
  MTag,
  property,
  paddock,
  date = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{property}{the name of the property that the cattle belong}

\item{date}{the date that the cattle were moved in date format, the default is today's date}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the information has been successfully updated
}
\description{
This function allows individuals or groups of cattle to be allocated to paddocks in the DataMuster database via the DataMuster website app
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
